add single store evenet handler, add baseNode, add support content as a callback
