+++ lib/customelements/ElementCreator.js
@@ -1,21 +1,5 @@
 (() => {
 	/**
-	 * Element Div proptotype
-	 */
-	HTMLElement.prototype.Div = function (props, content) {
-		const el = Div(props, content)
-		this.appendChild(el);
-		return el;
-	}
-	/**
-	 * Element Span prototype
-	 */
-	HTMLElement.prototype.Span = function (props, content) {
-		const el = Span(props, content);
-		this.appendChild(el);
-		return el
-	}
-	/**
 	 * Create a element with properties
 	 * @param {String} name Tag Name
 	 * @param {Object} props Propeperties
@@ -31,16 +15,26 @@
 			} else if (prop === 'classList'){
 				el.classList.add(...propVal);
 			} else if(prop === 'content') {
-				if (typeof propVal === 'string') {
-					el.innerHTML = propVal;
-				} else {
-					el.appendChild(propVal)
-				}
+				_addConetent(propVal, el);
 			} else {
 				el[prop] = propVal;
 			}
 		})
 		return el;
+	}
+	/**
+	 * Add content to a element
+	 * @param {*} content
+	 * @param {HTMLElement} el element to append content
+	 */
+	function _addConetent(content, el) {
+		if (typeof content === 'string') {
+			el.innerHTML = content;
+		} else if (typeof content === 'function') {
+			el.appendChild(content())
+		} else {
+			el.appendChild(contet)
+		}
 	}
 	/**
 	 * Create a div and return it
@@ -72,6 +66,7 @@
 		props.content = content ? content : '';
 		return props
 	}
+
 
 	module.exports = { HTMLElementCreator, Div, Span }
 })()
