+++ lib/index.js
@@ -2,7 +2,8 @@ const { MetaComponent } = require('./metaviews/meta-component');
 const { MetaContainer } = require('./metaviews/meta-container');
 const { MetaShadowComponent } = require('./metaviews/meta-shadow-component');
 const Store = require('./store');
-const { HTMLElementCreator, Div } = require('./ElementCreator');
+const { HTMLElementCreator, Div, Span } = require('./customelements/ElementCreator');
+require('./customelements');
 
 module.exports = {
 	MetaComponent,
@@ -10,5 +11,6 @@ module.exports = {
 	MetaShadowComponent,
 	Store,
 	HTMLElementCreator,
-	Div
+	Div,
+	Span
 };
